Pulled the deprecated Mutator and associated method into a sub-interface of WorkflowAction
This should be a find-and-replace migration, and allows `WorkflowAction` to require
implementors to override `apply()`. The new sub-interface is completely deprecated.
